requirements.txt
Django==1.3
BeautifulSoup==3.2.0
django-localeurl==1.4
django-rosetta==0.6.2
https://github.com/toastdriven/django-tastypie/tarball/master
django-transmeta==0.6.2
feedparser==5.0.1
gdata==2.0.14
microdata==0.3.0
rdfextras==0.1
rdflib==3.1.0
https://github.com/edsu/rdflib-microdata/tarball/master
simplejson==2.1.2
mimeparse==0.1.3
html5lib==0.90
python-dateutil==1.5
lxml
PyYAML
python_digest
django_mobile==0.2.1
django-haystack==1.2.4
Whoosh==2.2.2
